BOCM-288 GET endpoint for /more-appeals done and tested
